master: middleware test

+++ routes/api.php
@@ -30,9 +30,9 @@ Route::group(['prefix' => 'v1'], function () {
         Route::group(['prefix' => 'contact'], function () {
             Route::get('/', [ContactController::class, 'index'])->name('contact_index');
             Route::post('/', [ContactController::class, 'store'])->name('contact_store');
-            Route::get('/{contact}', [ContactController::class, 'show'])->name('contact_show');
-            Route::patch('/{contact}', [ContactController::class, 'update'])->name('contact_update');
-            Route::delete('/{contact}', [ContactController::class, 'destroy'])->name('contact_destroy');
+            Route::get('/{contact}', [ContactController::class, 'show'])->middleware('contactOwner')->name('contact_show');
+            Route::patch('/{contact}', [ContactController::class, 'update'])->middleware('contactOwner')->name('contact_update');
+            Route::delete('/{contact}', [ContactController::class, 'destroy'])->middleware('contactOwner')->name('contact_destroy');
         });
     });
